remove installer
