--- crates/common/precompiles/src/installer.rs
@@ -1,102 +1,0 @@
-use alloy_evm::precompiles::PrecompilesMap;
-use base_common_chains::BaseUpgrade;
-
-use crate::{ActivationRegistryPrecompile, BasePrecompileSpec, BasePrecompiles};
-
-/// Installs the full Base precompile set for a given spec.
-#[derive(Debug, Clone, Copy)]
-pub struct BasePrecompileInstaller<S = BaseUpgrade> {
-    /// Spec used to select the Base precompile set.
-    spec: S,
-}
-
-impl<S: BasePrecompileSpec> BasePrecompileInstaller<S> {
-    /// Creates a new installer for the given spec.
-    pub const fn new(spec: S) -> Self {
-        Self { spec }
-    }
-
-    /// Returns the spec used by this installer.
-    pub const fn spec(&self) -> S {
-        self.spec
-    }
-
-    /// Builds a [`PrecompilesMap`] with all Base precompiles installed.
-    pub fn install(self) -> PrecompilesMap {
-        let mut precompiles =
-            PrecompilesMap::from_static(BasePrecompiles::new_with_spec(self.spec).precompiles());
-        self.install_into(&mut precompiles);
-        precompiles
-    }
-
-    /// Installs Base-specific dynamic precompiles into an existing [`PrecompilesMap`].
-    pub fn install_into(self, precompiles: &mut PrecompilesMap) {
-        if self.spec.upgrade() >= BaseUpgrade::Beryl {
-            crate::TokenFactoryPrecompile::install(precompiles);
-            crate::B20TokenPrecompile::install(precompiles);
-            crate::PolicyRegistryEvm::install(precompiles);
-            ActivationRegistryPrecompile::install(precompiles);
-        }
-    }
-}
-
-impl<S: BasePrecompileSpec> Default for BasePrecompileInstaller<S> {
-    fn default() -> Self {
-        Self::new(S::default_precompile_spec())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use alloy_primitives::{Address, B256};
-    use revm::precompile::{bn254, secp256r1};
-    use rstest::rstest;
-
-    use super::*;
-    use crate::{ActivationRegistry, TokenFactory, TokenVariant};
-
-    #[test]
-    fn installer_preserves_base_precompile_set() {
-        let precompiles = BasePrecompileInstaller::new(BaseUpgrade::Jovian).install();
-
-        assert!(precompiles.get(&bn254::pair::ADDRESS).is_some());
-        assert!(precompiles.get(secp256r1::P256VERIFY.address()).is_some());
-    }
-
-    #[test]
-    fn default_installer_uses_default_precompile_spec() {
-        let installer = BasePrecompileInstaller::new(BaseUpgrade::LATEST);
-
-        assert_eq!(installer.spec(), BaseUpgrade::LATEST);
-    }
-
-    #[rstest]
-    #[case::azul(BaseUpgrade::Azul, false)]
-    #[case::beryl(BaseUpgrade::Beryl, true)]
-    fn installer_routes_b20_precompiles_by_fork(#[case] spec: BaseUpgrade, #[case] expected: bool) {
-        let precompiles = BasePrecompileInstaller::new(spec).install();
-        let (token, _) = TokenVariant::B20.compute_address(
-            Address::repeat_byte(0x11),
-            18,
-            B256::repeat_byte(0x22),
-        );
-
-        assert_eq!(precompiles.get(&TokenFactory::ADDRESS).is_some(), expected);
-        assert_eq!(precompiles.get(&token).is_some(), expected);
-        assert!(precompiles.get(&Address::repeat_byte(0x42)).is_none());
-    }
-
-    #[test]
-    fn activation_registry_is_not_installed_before_beryl() {
-        let precompiles = BasePrecompileInstaller::new(BaseUpgrade::Azul).install();
-
-        assert!(precompiles.get(&ActivationRegistry::ADDRESS).is_none());
-    }
-
-    #[test]
-    fn activation_registry_is_installed_at_beryl() {
-        let precompiles = BasePrecompileInstaller::new(BaseUpgrade::Beryl).install();
-
-        assert!(precompiles.get(&ActivationRegistry::ADDRESS).is_some());
-    }
-}
