+++ crates/common/precompiles/src/lib.rs
@@ -10,9 +10,6 @@ mod macros;
 mod provider;
 pub use provider::BasePrecompiles;
 
-mod installer;
-pub use installer::BasePrecompileInstaller;
-
 mod spec;
 pub use spec::BasePrecompileSpec;
